# lcm/ns/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "ZTE vManager Service rest API"
    },
    "basePath": "/openoapi/nslcm/v1",
...
    "paths": {
...
        "/ns/{ns_instance_id}/terminate": {
            "post": {
                "tags": [
                    "ns lcm"
                ],
                "summary": "ns terminate",
                "description": "ns terminate",
                "operationId": "ns_terminate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "required": true,
                        "type": "string",
                        "description": "Identifier of the NS instance.",
                        "name": "ns_instance_id",
                        "in": "path"
                    },
                    {
                        "in": "body",
                        "name": "NsTerminateRequest",
                        "description": "NsTerminateRequest",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NsTerminateRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/NsTerminateResponse"
                        }
                    },
                    "500": {
                        "description": "the url is invalid"
                    }
                }
            }
        },
        "/ns/{ns_instance_id}/heal": {
            "post": {
                "tags": [
                    "ns heal"
                ],
                "summary": "ns heal",
                "description": "ns heal",
                "operationId": "ns_heal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "required": true,
                        "type": "string",
                        "description": "Identifier of the NS instance.",
                        "name": "ns_instance_id",
                        "in": "path"
                    },
                    {
                        "in": "body",
                        "name": "healVnfData",
                        "description": "healVnfData",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/healVnfDataRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/healVnfDataResponse"
                        }
                    },
                    "500": {
                        "description": "the url is invalid"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "NsTerminateRequest": {
            "type": "object",
            "properties": {
                "terminationType": {
                    "type": "string"
                },
                "gracefulTerminationTimeout": {
                    "type": "string"
                }
            }
        },
        "NsTerminateResponse": {
            "type": "object",
            "properties": {
                "jobID": {
                    "type": "string"
                }
            }
        },
        "healVnfDataRequest": {
            "type": "object",
            "properties": {
                "vnfInstanceId": {
                    "type": "string"
                },
                "cause": {
                    "type": "string"
                },
                "additionalParams": {
                    "type": "object",
                    "properties": {
                        "action": {
                            "type": "string"
                        },
                        "actionvminfo": {
                            "type": "object",
                            "properties": {
                                "vmid": {
                                    "type": "string"
                                },
                                "vmname": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "healVnfDataResponse": {
            "type": "object",
            "properties": {
                "jobId": {
                    "type": "string"
                }
            }
        },
...
    }
}